fata mapper

+++ spring/src/main/java/com/programacao/web/fatec/apifatec/domain/produto/ProdutoService.java
@@ -11,39 +11,41 @@ import com.programacao.web.fatec.apifatec.entitines.Produto;
 public class ProdutoService {
 	@Autowired
 	private ProdutoRepository produtoRepository;
-	
-	public List <Produto> ListarProdutos(){
+
+	public List<Produto> listarProdutos() {
 		return produtoRepository.findAll();
 	}
-	
+
 	public Produto encontrarProdutoPorId(Long id) {
 		return produtoRepository.findById(id).orElse(null);
 	}
-	
-	public Produto SalvarProduto (Produto produto) {
-		return produtoRepository.save(produto);		
-	}
-	
-	public void deletarProduto (Long id) {
-		Produto produto = encontrarProdutoPorId(id);
-		if (produto == null)
-		{
-			
-		}
-		
-		produtoRepository.deleteById(id);		
-	}
-	public Produto atualizarProduto(Long id, Produto produto) {
-		Produto produtoCadastradoProduto = encontrarProdutoPorId(id);
-		
-		if (produtoCadastradoProduto==null)
-		{
-			return null;
-		}else {
-			produtoCadastradoProduto.setNome(produto.getNome());
-		}
-		
-		
+
+	public Produto salvarProduto(Produto Produto) {
+		return produtoRepository.save(Produto);
 	}
 
+	public void deletarProduto(Long id) {
+		Produto Produto = encontrarProdutoPorId(id);
+		
+		if(Produto == null)
+			throw new IllegalArgumentException("Produto nao existe");
+		
+		
+		produtoRepository.deleteById(id);
+	}
+	
+	public Produto atualizarProduto(Long id, Produto produto) {
+		Produto ProdutoCadastrado = encontrarProdutoPorId(id);
+		
+		if (ProdutoCadastrado == null)
+		{
+			return null;
+		} else {
+			ProdutoCadastrado.setProduto(produto.getProduto());
+			ProdutoCadastrado.setPreco(produto.getPreco());
+			ProdutoCadastrado.setQuantidade(produto.getQuantidade());
+			ProdutoCadastrado.setAtivo(produto.isAtivo());
+			return produtoRepository.save(ProdutoCadastrado);
+		}
+	}
 }
